ARM: drop warning about return_address not using unwind tables
applied patch

The warning was introduced in 2009 (commit 4bf1fa5a34aa ([ARM] 5613/1:
implement CALLER_ADDRESSx)). The only result is that CALLER_ADDRESSx for
x > 1 returns NULL which doesn't do much harm.

The drawback of implementing the use of unwind tables for the kernel
tracers is that much of the unwinder code would need to be marked as not
traceable.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
---
 arch/arm/kernel/return_address.c | 4 ----
 1 file changed, 4 deletions(-)